organize files (folders and lists) and fix errors, cleaning code
