+++ app/src/components/dashboard/advantages/AdvantageHome.js
@@ -1,7 +1,7 @@
 import AdvantagesList from "./AdvantageList";
-import ADVANTAGES_LIST from "./advantage_list";
+import ADVANTAGES_LIST from "../../../lists/advantage_list";
 
-export default function List() {
+export default function AdvantageHome() {
   return (
     <div className="pb-20 pt-20 h-screen text-center ">
       <p className="pb-10 text-2xl opensans">ENTRE EM CONTACTO</p>
